modify delete variableCsot route to check if cost is present in some recipes

+++ routes/variableCosts.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const VariableCost = require("../models/VariableCost");
+const RecipeCost = require("../models/RecipeCost");
 
 const router = express.Router();
 
@@ -39,16 +40,42 @@ router.post("/update", async (req, res, next) => {
 
 router.delete("/delete", async (req, res, next) => {
   const { id } = req.body;
-  try {
-    const result = await VariableCost.destroy({
-      where: { id },
-      force: true,
-    });
-    res.json(result);
-  } catch (err) {
-    res.status(500);
-    res.json(err);
+  const recipesWithCost = await findVariableCostInRecipes(id);
+  if (recipesWithCost.length == 0) {
+    try {
+      const result = await VariableCost.destroy({
+        where: { id },
+        force: true,
+      });
+      res.json(result);
+    } catch (err) {
+      res.status(500);
+      res.json(err);
+    }
+  } else {
+    res.status(409)
+    res.json(
+      {
+        currently_used: true,
+        recipes: recipesWithCost
+      }
+    )
   }
 });
+
+const findVariableCostInRecipes = async (costId) => {
+  try {
+    const foundRecipesIds = [];
+    const foundRecipes = await RecipeCost.findAll({
+      where: { VariableCostId: costId },
+    });
+    foundRecipes.forEach((recipe) => {
+      foundRecipesIds.push(recipe.RecipeId);
+    });
+    return foundRecipesIds;
+  } catch (err) {
+    return err;
+  }
+};
 
 module.exports = router;
